4_al_backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
supabase==2.9.1
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.5.2
xgboost==2.1.1
python-dotenv==1.0.1
joblib==1.4.2
pytest==8.3.3
httpx==0.27.2
